Removed packages, dist and renamed test project
