+++ .gitignore
@@ -1,12 +1,5 @@
-/bin/Release
-*/obj/Debug
-*/obj/Release
-/*.suo
-/Loop54/bin/Debug
-/Loop54/obj/Debug
-/Loop54/bin/Release
-/Loop54/obj/Release
-/NUnitTests/bin/Debug
-/NUnitTests/obj/Debug
-/.vs/Loop54/v14/.suo
-/packages/NUnit.3.4.1
+*/bin/*
+*/obj/*
+*.suo
+.vs/*
+packages
